+++ src/alited.tcl
@@ -7,7 +7,7 @@
 # License: MIT.
 ###########################################################
 
-package provide alited 1.4.5.2  ;# for documentation (esp. for Ruff!)
+package provide alited 1.4.5.3  ;# for documentation (esp. for Ruff!)
 
 namespace eval alited {